feat(eval): calibrate OCR/visual agent-task corpus against TS baseline
Expand the capability-first local agent-task corpus with visual candidates,
OCR text-layer, and configured-command visual enrichment fusion tasks.
Share metric/acceptance helpers, remeasure the TypeScript 3.0.14 baseline,
and keep pure-Rust comparison on measured presence floors under publish freeze.
