+++ docs/specs/agent-task-corpus/tasks/ocr-text-layer.json
@@ -1,0 +1,27 @@
+{
+  "schemaVersion": 1,
+  "id": "ocr-text-layer-mock-provider",
+  "title": "OCR text layer appears for scanned/sparse page with configured command provider",
+  "class": "ocr_scanned_page",
+  "status": "active",
+  "tool": "read_pdf",
+  "fixture": "test/fixtures/differential/v3014-visual-v1.pdf",
+  "env": {
+    "MCP_PDF_OCR_COMMAND": "{node}",
+    "MCP_PDF_OCR_ARGS_JSON": "[\"{root}/test/fixtures/mock-ocr-provider.mjs\",\"{input}\",\"{page}\",\"{languages}\"]"
+  },
+  "clearEnv": ["MCP_PDF_OCR_PRESET"],
+  "input": {
+    "sources": [{ "path": "test/fixtures/differential/v3014-visual-v1.pdf", "pages": [1] }],
+    "auto": false,
+    "include_page_count": true,
+    "include_document_map": true,
+    "include_ocr_text_layer": true
+  },
+  "acceptance": {
+    "resultSuccess": true,
+    "requireOcrTextLayer": true,
+    "minOcrTextChars": 5
+  },
+  "contractIds": ["semantic-ocr-text-layer"]
+}
