+++ docs/specs/agent-task-corpus/tasks/visual-candidates.json
@@ -1,0 +1,22 @@
+{
+  "schemaVersion": 1,
+  "id": "visual-candidates-provider-independent",
+  "title": "Visual region candidates remain discoverable without a provider",
+  "class": "visual_candidates",
+  "status": "active",
+  "tool": "read_pdf",
+  "fixture": "test/fixtures/differential/v3014-visual-candidate-v1.pdf",
+  "input": {
+    "sources": [{ "path": "test/fixtures/differential/v3014-visual-candidate-v1.pdf", "pages": [1] }],
+    "auto": false,
+    "include_page_count": true,
+    "include_document_map": true,
+    "include_visual_enrichments": true
+  },
+  "acceptance": {
+    "resultSuccess": true,
+    "minVisualCandidates": 1,
+    "requireDocumentMap": true
+  },
+  "contractIds": ["semantic-visual-candidates", "semantic-visual-region-analysis"]
+}
