+++ docs/specs/agent-task-corpus/tasks/visual-enrichment-fusion.json
@@ -1,0 +1,28 @@
+{
+  "schemaVersion": 1,
+  "id": "visual-enrichment-document-map-fusion",
+  "title": "Configured-command visual enrichment fuses into Document Map",
+  "class": "visual_enrichment",
+  "status": "active",
+  "tool": "read_pdf",
+  "fixture": "test/fixtures/differential/v3014-visual-candidate-v1.pdf",
+  "env": {
+    "MCP_PDF_REGION_ANALYSIS_COMMAND": "{node}",
+    "MCP_PDF_REGION_ANALYSIS_ARGS_JSON": "[\"{root}/scripts/differential/reference-visual-fusion-provider.ts\",\"{input}\",\"{page}\",\"{region_id}\",\"{evidence_id}\",\"{languages}\",\"success\"]"
+  },
+  "clearEnv": ["MCP_PDF_REGION_ANALYSIS_HTTP_URL", "MCP_PDF_REGION_ANALYSIS_PRESET"],
+  "input": {
+    "sources": [{ "path": "test/fixtures/differential/v3014-visual-candidate-v1.pdf", "pages": [1] }],
+    "auto": false,
+    "include_page_count": true,
+    "include_document_map": true,
+    "include_visual_enrichments": true
+  },
+  "acceptance": {
+    "resultSuccess": true,
+    "minVisualCandidates": 1,
+    "minVisualEnrichments": 1,
+    "requireDocumentMap": true
+  },
+  "contractIds": ["semantic-visual-region-analysis", "semantic-visual-candidates"]
+}
