+++ docs/specs/semantic-contracts/semantic-ocr-text-layer.json
@@ -1,0 +1,30 @@
+{
+  "schemaVersion": 1,
+  "id": "semantic-ocr-text-layer",
+  "title": "OCR text layer remains task-usable for scanned/sparse pages",
+  "status": "active",
+  "layer": "semantic",
+  "capability": "ocr_text_layer",
+  "tools": ["read_pdf", "search_pdf"],
+  "matrixPaths": ["tools.read_pdf.include_ocr_text_layer", "tools.search_pdf.include_ocr_text_layer"],
+  "mustHold": [
+    "When an OCR command provider is configured, include_ocr_text_layer produces usable OCR text evidence on admitted fixtures",
+    "OCR evidence retains page association sufficient for citation",
+    "Not-configured OCR remains fail-soft rather than inventing a layer"
+  ],
+  "mayDiffer": [
+    "Exact word geometry and confidence numeric representation",
+    "Whitespace and token boundaries within semantic equivalence"
+  ],
+  "failClosed": [
+    "Provider failure does not invent OCR full-page text as a success path without warnings"
+  ],
+  "evidence": {
+    "kind": "task-eval",
+    "suites": [
+      "scripts/run-agent-task-smoke.ts",
+      "scripts/run-agent-task-eval.ts",
+      "scripts/differential/check-v3014-read-ocr-differential.ts"
+    ]
+  }
+}
