+++ docs/specs/semantic-contracts/semantic-visual-candidates.json
@@ -1,0 +1,30 @@
+{
+  "schemaVersion": 1,
+  "id": "semantic-visual-candidates",
+  "title": "Visual region candidates remain task-usable without requiring PDF.js equality",
+  "status": "active",
+  "layer": "semantic",
+  "capability": "visual_candidates",
+  "tools": ["read_pdf"],
+  "matrixPaths": ["tools.read_pdf.include_visual_enrichments", "tools.read_pdf.include_document_map"],
+  "mustHold": [
+    "Admitted visual fixtures expose visual enrichment candidates when include_visual_enrichments is enabled",
+    "Candidates remain available even when no analysis provider is configured",
+    "Document Map can route candidate presence for agent follow-up"
+  ],
+  "mayDiffer": [
+    "Exact candidate IDs and sub-pixel bounding boxes",
+    "Ordering within an equivalence class of candidates"
+  ],
+  "failClosed": [
+    "Provider absence does not invent enrichment payloads"
+  ],
+  "evidence": {
+    "kind": "task-eval",
+    "suites": [
+      "scripts/run-agent-task-smoke.ts",
+      "scripts/run-agent-task-eval.ts",
+      "scripts/differential/check-v3014-visual-candidate-differential.ts"
+    ]
+  }
+}
