+++ scripts/run-agent-task-eval.ts
@@ -2,14 +2,25 @@
 /**
  * Capability-first agent-task evaluation (ADR-0005).
  *
- * - Measures semantic metrics on the same local corpus for TS baseline and
- *   pure-Rust candidate.
- * - Thresholds come from measured TS baseline metrics, never invented %.
- * - Exact PDF.js JSON equality is intentionally not required.
+ * Measures semantic metrics on the same local corpus for TS baseline and
+ * pure-Rust candidate. Thresholds come from measured TS baseline metrics.
  */
-import { spawn, spawnSync, type ChildProcessWithoutNullStreams } from 'node:child_process';
+import { spawnSync } from 'node:child_process';
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 import { dirname, join } from 'node:path';
+import {
+  callMcpTool,
+  evaluateAcceptance,
+  extractMetrics,
+  loadTasks,
+  publicPayload,
+  resolveInput,
+  resolveTaskEnv,
+  type Metrics,
+  type RuntimeId,
+  type Task,
+  type TaskMeasurement,
+} from './agent-task-shared.ts';
 
 const root = join(import.meta.dirname, '..');
 const corpusDir = join(root, 'docs/specs/agent-task-corpus');
@@ -17,44 +28,6 @@ const manifestPath = join(corpusDir, 'manifest.json');
 const baselinePath = join(corpusDir, 'baselines/typescript-v3.0.14.local.json');
 const rustServerPath = join(root, 'target/release/pdf-reader-mcp-server');
 const tsServerPath = join(root, 'dist/index.js');
-
-type Json = null | boolean | number | string | Json[] | { [key: string]: Json };
-type RuntimeId = 'typescript' | 'pure-rust';
-
-type Task = {
-  id: string;
-  class?: string;
-  tool: string;
-  fixture?: string;
-  input: Record<string, unknown>;
-  acceptance: Record<string, unknown>;
-  contractIds?: string[];
-};
-
-type Metrics = {
-  success: boolean;
-  isError: boolean;
-  fullTextChars: number;
-  pageCount: number | null;
-  tableCount: number;
-  matchCount: number;
-  firstMatchPage: number | null;
-  firstTablePage: number | null;
-  outlineCount: number;
-  formFieldCount: number;
-  annotationCount: number;
-  warningCount: number;
-  inventedFullTextRisk: boolean;
-};
-
-type TaskMeasurement = {
-  taskId: string;
-  class?: string;
-  runtime: RuntimeId;
-  metrics: Metrics;
-  acceptancePass: boolean;
-  acceptanceFailures: string[];
-};
 
 const args = process.argv.slice(2);
 const writeBaseline = args.includes('--write-baseline');
@@ -69,8 +42,6 @@ const runtimes: RuntimeId[] = args.includes('--runtime=typescript')
 
 const manifest = JSON.parse(readFileSync(manifestPath, 'utf8')) as {
   taskFiles: string[];
-  baselineRuntime?: string;
-  candidateRuntime?: string;
   calibration?: { inventedNumericThresholdsForbidden?: boolean };
 };
 
@@ -97,302 +68,42 @@ const ensureTsServer = () => {
   process.exit(1);
 };
 
-const callTool = async (
-  runtime: RuntimeId,
-  tool: string,
-  toolArgs: Record<string, unknown>
-): Promise<Record<string, unknown>> => {
-  const child =
-    runtime === 'pure-rust'
-      ? (spawn(rustServerPath, [], {
-          cwd: root,
-          stdio: ['pipe', 'pipe', 'pipe'],
-          env: { ...process.env, MCP_TRANSPORT: 'stdio', PDF_READER_ENGINE_MODE: 'pure-rust' },
-        }) as ChildProcessWithoutNullStreams)
-      : (spawn(process.execPath, [tsServerPath], {
-          cwd: root,
-          stdio: ['pipe', 'pipe', 'pipe'],
-          env: {
-            ...process.env,
-            MCP_TRANSPORT: 'stdio',
-            // Explicitly leave pure-rust opt-in unset for TS LKG path.
-            PDF_READER_ENGINE_MODE: '',
-          },
-        }) as ChildProcessWithoutNullStreams);
-
-  let buffer = '';
-  let stderr = '';
-  const pending = new Map<number, (value: Record<string, unknown>) => void>();
-  child.stderr.on('data', (chunk) => {
-    stderr += chunk.toString();
-  });
-  child.stdout.on('data', (chunk) => {
-    buffer += chunk.toString();
-    while (buffer.includes('\n')) {
-      const index = buffer.indexOf('\n');
-      const line = buffer.slice(0, index).trim();
-      buffer = buffer.slice(index + 1);
-      if (!line) continue;
-      try {
-        const response = JSON.parse(line) as Record<string, unknown>;
-        const id = Number(response.id);
-        const resolver = pending.get(id);
-        if (resolver) {
-          pending.delete(id);
-          resolver(response);
-        }
-      } catch {
-        // ignore non-JSON noise
-      }
-    }
-  });
-
-  const request = (id: number, method: string, params: Record<string, unknown>) =>
-    new Promise<Record<string, unknown>>((resolve, reject) => {
-      const timer = setTimeout(
-        () => reject(new Error(`timeout ${runtime} ${method}: ${stderr.slice(-2000)}`)),
-        45000
-      );
-      pending.set(id, (value) => {
-        clearTimeout(timer);
-        resolve(value);
-      });
-      child.stdin.write(`${JSON.stringify({ jsonrpc: '2.0', id, method, params })}\n`);
-    });
-
-  try {
-    await request(1, 'initialize', {
-      protocolVersion: '2024-11-05',
-      capabilities: {},
-      clientInfo: { name: 'agent-task-eval', version: '1' },
-    });
-    child.stdin.write(
-      `${JSON.stringify({ jsonrpc: '2.0', method: 'notifications/initialized' })}\n`
-    );
-    return await request(2, 'tools/call', { name: tool, arguments: toolArgs });
-  } finally {
-    child.kill('SIGTERM');
-  }
-};
-
-const publicPayload = (response: Record<string, unknown>): Record<string, unknown> => {
-  const result = response.result as Record<string, unknown> | undefined;
-  if (!result) throw new Error(`missing result: ${JSON.stringify(response).slice(0, 500)}`);
-  if (Array.isArray(result.content)) {
-    const textPart = result.content.find(
-      (entry) => entry && typeof entry === 'object' && (entry as { type?: string }).type === 'text'
-    ) as { text?: string } | undefined;
-    if (textPart?.text) return JSON.parse(textPart.text) as Record<string, unknown>;
-  }
-  if (result.structuredContent && typeof result.structuredContent === 'object') {
-    return result.structuredContent as Record<string, unknown>;
-  }
-  return result;
-};
-
-const asRecordArray = (value: unknown): Array<Record<string, unknown>> => {
-  if (!Array.isArray(value)) return [];
-  return value.filter((entry) => entry && typeof entry === 'object') as Array<
-    Record<string, unknown>
-  >;
-};
-
-const resolveInput = (input: Record<string, unknown>): Record<string, unknown> => {
-  const resolved = structuredClone(input);
-  if (Array.isArray(resolved.sources)) {
-    for (const source of resolved.sources as Array<Record<string, unknown>>) {
-      if (typeof source.path === 'string' && !source.path.startsWith('/')) {
-        source.path = join(root, source.path);
-      }
-    }
-  }
-  return resolved;
-};
-
-const extractMetrics = (
-  response: Record<string, unknown>,
-  payload: Record<string, unknown>
-): Metrics => {
-  const result = (response.result as Record<string, unknown> | undefined) ?? {};
-  const isError = result.isError === true;
-  const results = asRecordArray(payload.results);
-  const first = (results[0] ?? payload) as Record<string, unknown>;
-  const data =
-    first.data && typeof first.data === 'object'
-      ? (first.data as Record<string, unknown>)
-      : first;
-
-  const pageTexts = asRecordArray(data.page_texts ?? first.page_texts);
-  const joinedPageText = pageTexts
-    .map((entry) => (typeof entry.text === 'string' ? entry.text : ''))
-    .join('\n');
-  const fullText =
-    (typeof data.full_text === 'string' ? data.full_text : undefined) ??
-    (typeof first.full_text === 'string' ? first.full_text : undefined) ??
-    (typeof data.text === 'string' ? data.text : undefined) ??
-    (typeof first.text === 'string' ? first.text : undefined) ??
-    joinedPageText ??
-    '';
-
-  const pageCountRaw =
-    data.page_count ?? first.page_count ?? data.num_pages ?? first.num_pages ?? null;
-  const pageCount =
-    typeof pageCountRaw === 'number'
-      ? pageCountRaw
-      : typeof pageCountRaw === 'string' && pageCountRaw.trim()
-        ? Number(pageCountRaw)
-        : null;
-
-  const tablesTop = asRecordArray(data.tables ?? first.tables);
-  const tableInfo = asRecordArray(data.table_info ?? first.table_info);
-  const elementTables = asRecordArray(data.elements ?? first.elements).filter(
-    (entry) => entry.kind === 'table' || entry.type === 'table'
-  );
-  const tables = tablesTop.length ? tablesTop : tableInfo.length ? tableInfo : elementTables;
-  const firstTablePage =
-    tables[0] && typeof tables[0].page === 'number'
-      ? Number(tables[0].page)
-      : tables[0]
-        ? 1
-        : null;
-
-  const matchesTop = asRecordArray(
-    (first.matches as unknown) ?? (data.matches as unknown) ?? (payload.matches as unknown)
-  );
-  const nestedMatches = results.flatMap((entry) => asRecordArray(entry.matches));
-  const matches = matchesTop.length ? matchesTop : nestedMatches;
-  const firstMatchPage =
-    matches[0] && typeof matches[0].page === 'number' ? Number(matches[0].page) : null;
-
-  const outline = asRecordArray(
-    data.outline ?? first.outline ?? data.outlines ?? first.outlines
-  );
-  const formFields = asRecordArray(
-    data.form_fields ?? first.form_fields ?? data.fields ?? first.fields
-  );
-  const annotations = asRecordArray(data.annotations ?? first.annotations);
-  const warnings = asRecordArray(data.warnings ?? first.warnings ?? payload.warnings);
-
-  const success = !isError && (first.success === true || first.success === undefined);
-
-  return {
-    success,
-    isError,
-    fullTextChars: fullText.length,
-    pageCount: Number.isFinite(pageCount as number) ? (pageCount as number) : null,
-    tableCount: tables.length,
-    matchCount: matches.length,
-    firstMatchPage,
-    firstTablePage,
-    outlineCount: outline.length,
-    formFieldCount: formFields.length,
-    annotationCount: annotations.length,
-    warningCount: warnings.length,
-    // Only meaningful for fail-closed invalid-page tasks; computed during acceptance.
-    inventedFullTextRisk: false,
-  };
-};
-
-const evaluateAcceptance = (
-  acceptance: Record<string, unknown>,
-  metrics: Metrics
-): { pass: boolean; failures: string[] } => {
-  const failures: string[] = [];
-  if (acceptance.resultSuccess === true && !metrics.success) {
-    failures.push('resultSuccess required');
-  }
-  if (typeof acceptance.minFullTextChars === 'number') {
-    if (metrics.fullTextChars < acceptance.minFullTextChars) {
-      failures.push(
-        `minFullTextChars ${acceptance.minFullTextChars}, got ${metrics.fullTextChars}`
-      );
-    }
-  }
-  if (typeof acceptance.requirePageCountAtLeast === 'number') {
-    if ((metrics.pageCount ?? 0) < acceptance.requirePageCountAtLeast) {
-      failures.push(
-        `requirePageCountAtLeast ${acceptance.requirePageCountAtLeast}, got ${metrics.pageCount}`
-      );
-    }
-  }
-  if (typeof acceptance.minTables === 'number') {
-    if (metrics.tableCount < acceptance.minTables) {
-      failures.push(`minTables ${acceptance.minTables}, got ${metrics.tableCount}`);
-    }
-  }
-  if (acceptance.requireTablePage === 1 && metrics.tableCount > 0) {
-    if ((metrics.firstTablePage ?? 0) < 1) failures.push('table page invalid');
-  }
-  if (typeof acceptance.minMatches === 'number') {
-    if (metrics.matchCount < acceptance.minMatches) {
-      failures.push(`minMatches ${acceptance.minMatches}, got ${metrics.matchCount}`);
-    }
-  }
-  if (typeof acceptance.requireMatchPageAtLeast === 'number' && metrics.matchCount > 0) {
-    if ((metrics.firstMatchPage ?? 0) < acceptance.requireMatchPageAtLeast) {
-      failures.push('match page missing/invalid');
-    }
-  }
-  if (typeof acceptance.minOutlineItems === 'number') {
-    if (metrics.outlineCount < acceptance.minOutlineItems) {
-      failures.push(`minOutlineItems ${acceptance.minOutlineItems}, got ${metrics.outlineCount}`);
-    }
-  }
-  if (typeof acceptance.minFormFields === 'number') {
-    if (metrics.formFieldCount < acceptance.minFormFields) {
-      failures.push(`minFormFields ${acceptance.minFormFields}, got ${metrics.formFieldCount}`);
-    }
-  }
-  if (typeof acceptance.minAnnotations === 'number') {
-    if (metrics.annotationCount < acceptance.minAnnotations) {
-      failures.push(
-        `minAnnotations ${acceptance.minAnnotations}, got ${metrics.annotationCount}`
-      );
-    }
-  }
-  if (acceptance.forbidInventedFullText === true && metrics.fullTextChars > 200) {
-    metrics.inventedFullTextRisk = true;
-    failures.push('invalid page produced excessive text');
-  }
-  if (acceptance.requireWarningOrEmptyText === true) {
-    if (metrics.fullTextChars >= 40 && metrics.warningCount === 0) {
-      failures.push('invalid page invented substantial text without warnings');
-    }
-  }
-  if (acceptance.allowSuccessWithWarnings === true) {
-    // success or warning/empty-text path is acceptable; hard isError alone is not auto-fail
-    // unless resultSuccess was also required.
-  }
-  return { pass: failures.length === 0, failures };
-};
-
-const compareMetricFloor = (
-  taskId: string,
-  metric: keyof Metrics,
-  baseline: number,
-  candidate: number,
-  mode: 'gte' | 'lte' = 'gte'
-): string | null => {
-  if (mode === 'gte' && candidate + 1e-9 < baseline) {
-    return `${taskId}: ${metric} below TS baseline (${candidate} < ${baseline})`;
-  }
-  if (mode === 'lte' && candidate - 1e-9 > baseline) {
-    return `${taskId}: ${metric} above TS baseline ceiling (${candidate} > ${baseline})`;
-  }
-  return null;
-};
-
-const loadTasks = (): Task[] =>
-  manifest.taskFiles.map((rel) => {
-    const task = JSON.parse(readFileSync(join(corpusDir, rel), 'utf8')) as Task;
-    return task;
-  });
-
 const measureRuntime = async (runtime: RuntimeId, tasks: Task[]): Promise<TaskMeasurement[]> => {
   const out: TaskMeasurement[] = [];
   for (const task of tasks) {
     try {
-      const response = await callTool(runtime, task.tool, resolveInput(task.input));
+      const baseEnv =
+        runtime === 'pure-rust'
+          ? {
+              ...process.env,
+              MCP_TRANSPORT: 'stdio',
+              PDF_READER_ENGINE_MODE: 'pure-rust',
+            }
+          : {
+              ...process.env,
+              MCP_TRANSPORT: 'stdio',
+              PDF_READER_ENGINE_MODE: '',
+            };
+      const env = resolveTaskEnv(task, root, baseEnv);
+      const response =
+        runtime === 'pure-rust'
+          ? await callMcpTool({
+              command: rustServerPath,
+              env,
+              tool: task.tool,
+              toolArgs: resolveInput(task.input, root),
+              cwd: root,
+              timeoutMs: 60_000,
+            })
+          : await callMcpTool({
+              command: process.execPath,
+              args: [tsServerPath],
+              env,
+              tool: task.tool,
+              toolArgs: resolveInput(task.input, root),
+              cwd: root,
+              timeoutMs: 60_000,
+            });
       const payload = publicPayload(response);
       const metrics = extractMetrics(response, payload);
       const acceptance = evaluateAcceptance(task.acceptance, metrics);
@@ -423,6 +134,10 @@ const measureRuntime = async (runtime: RuntimeId, tasks: Task[]): Promise<TaskMe
           annotationCount: 0,
           warningCount: 0,
           inventedFullTextRisk: false,
+          visualCandidateCount: 0,
+          visualEnrichmentCount: 0,
+          ocrTextChars: 0,
+          hasDocumentMap: false,
         },
         acceptancePass: false,
         acceptanceFailures: [error instanceof Error ? error.message : String(error)],
@@ -432,7 +147,19 @@ const measureRuntime = async (runtime: RuntimeId, tasks: Task[]): Promise<TaskMe
   return out;
 };
 
-const tasks = loadTasks();
+const compareMetricFloor = (
+  taskId: string,
+  metric: keyof Metrics,
+  baseline: number,
+  candidate: number
+): string | null => {
+  if (candidate + 1e-9 < baseline) {
+    return `${taskId}: ${metric} below TS baseline (${candidate} < ${baseline})`;
+  }
+  return null;
+};
+
+const tasks = loadTasks(root, manifest.taskFiles);
 if (runtimes.includes('pure-rust')) ensureRustServer();
 if (runtimes.includes('typescript')) ensureTsServer();
 
@@ -458,6 +185,9 @@ if (writeBaseline) {
   }
   if (ts.some((item) => !item.acceptancePass)) {
     console.error('[agent-task-eval] refusing to write baseline with failing TS acceptance');
+    for (const item of ts.filter((entry) => !entry.acceptancePass)) {
+      console.error(`${item.taskId}: ${item.acceptanceFailures.join('; ')}`);
+    }
     process.exit(1);
   }
   const baseline = {
@@ -466,7 +196,7 @@ if (writeBaseline) {
     runtime: 'typescript-v3.0.14',
     measuredAt: new Date().toISOString(),
     method: 'scripts/run-agent-task-eval.ts --write-baseline',
-    note: 'Measured local-fixture metrics. Thresholds must not be invented above these floors.',
+    note: 'Measured local-fixture metrics including OCR/visual task classes. No invented % thresholds.',
     tasks: Object.fromEntries(
       ts.map((item) => [
         item.taskId,
@@ -494,14 +224,10 @@ if (compareToBaseline && existsSync(baselinePath)) {
       failures.push(`${item.taskId}: missing measured TS baseline entry`);
       continue;
     }
-    // Floors from measured TS baseline; representation differences allowed.
-    // Capability presence floors from measured TS baseline (no invented %).
-    // Only compare metrics that are material to the task class / non-zero capability proof.
-    // Exact count equality is intentionally not required under ADR-0005.
     const task = tasks.find((entry) => entry.id === item.taskId);
     const acceptance = task?.acceptance ?? {};
     const presence: Array<[keyof Metrics, number, number]> = [];
-    if (typeof acceptance.minFullTextChars === 'number' || (task?.class === 'extract_passage')) {
+    if (typeof acceptance.minFullTextChars === 'number' || task?.class === 'extract_passage') {
       presence.push(['fullTextChars', base.fullTextChars > 0 ? 1 : 0, item.metrics.fullTextChars]);
     }
     if (typeof acceptance.minTables === 'number' || base.tableCount > 0) {
@@ -514,20 +240,47 @@ if (compareToBaseline && existsSync(baselinePath)) {
       presence.push(['outlineCount', base.outlineCount > 0 ? 1 : 0, item.metrics.outlineCount]);
     }
     if (typeof acceptance.minFormFields === 'number' || base.formFieldCount > 0) {
-      presence.push(['formFieldCount', base.formFieldCount > 0 ? 1 : 0, item.metrics.formFieldCount]);
+      presence.push([
+        'formFieldCount',
+        base.formFieldCount > 0 ? 1 : 0,
+        item.metrics.formFieldCount,
+      ]);
     }
     if (typeof acceptance.minAnnotations === 'number' || base.annotationCount > 0) {
-      presence.push(['annotationCount', base.annotationCount > 0 ? 1 : 0, item.metrics.annotationCount]);
+      presence.push([
+        'annotationCount',
+        base.annotationCount > 0 ? 1 : 0,
+        item.metrics.annotationCount,
+      ]);
+    }
+    if (typeof acceptance.minVisualCandidates === 'number' || base.visualCandidateCount > 0) {
+      presence.push([
+        'visualCandidateCount',
+        base.visualCandidateCount > 0 ? 1 : 0,
+        item.metrics.visualCandidateCount,
+      ]);
+    }
+    if (typeof acceptance.minVisualEnrichments === 'number' || base.visualEnrichmentCount > 0) {
+      presence.push([
+        'visualEnrichmentCount',
+        base.visualEnrichmentCount > 0 ? 1 : 0,
+        item.metrics.visualEnrichmentCount,
+      ]);
+    }
+    if (
+      acceptance.requireOcrTextLayer === true ||
+      typeof acceptance.minOcrTextChars === 'number' ||
+      base.ocrTextChars > 0
+    ) {
+      presence.push(['ocrTextChars', base.ocrTextChars > 0 ? 1 : 0, item.metrics.ocrTextChars]);
     }
     for (const [metric, required, c] of presence) {
       if (required <= 0) continue;
-      const msg = compareMetricFloor(item.taskId, metric, required, c, 'gte');
+      const msg = compareMetricFloor(item.taskId, metric, required, c);
       if (msg) failures.push(`${msg} [TS measured presence]`);
     }
-    if (base.pageCount != null && base.pageCount >= 1 && acceptance.requirePageCountAtLeast != null) {
-      if ((item.metrics.pageCount ?? 0) < 1) {
-        failures.push(`${item.taskId}: pageCount missing while TS baseline had pageCount`);
-      }
+    if (acceptance.requireDocumentMap === true && base.hasDocumentMap && !item.metrics.hasDocumentMap) {
+      failures.push(`${item.taskId}: document_map missing while TS baseline had document_map`);
     }
     if (acceptance.forbidInventedFullText === true) {
       if (base.inventedFullTextRisk === false && item.metrics.fullTextChars > 200) {
@@ -536,9 +289,7 @@ if (compareToBaseline && existsSync(baselinePath)) {
     }
   }
 } else if (compareToBaseline && !existsSync(baselinePath) && !writeBaseline) {
-  failures.push(
-    `missing measured baseline at ${baselinePath}; run with --write-baseline first`
-  );
+  failures.push(`missing measured baseline at ${baselinePath}; run with --write-baseline first`);
 }
 
 const summary = {
